refactor(xaa): move flow-core types + capability-preflight to the SDK
Relocate the XAA flow-core types (executor seam, XAAFlowState/Step, config,
state-machine interface, DCR cache types, EMPTY_XAA_FLOW_STATE) and the
AS-compatibility preflight into sdk/src/xaa/state-machines/, exported from the
browser entry. The inspector's hosted resource-app wire types stay client-owned;
client types.ts + capability-preflight.ts become thin re-export shims so the
~20 UI importers are unchanged. Moved capability-preflight tests to the SDK.
